Create Bioinformatics KRAS Project
The RAS oncogenes are some of the most frequently mutated genes causing cancer. Research into this family of genes and proteins is biologically significant in combating cancer. In this research paper we show the biological and evolutionary similarity of KRAS proteins in humans compared to other species. We try to establish the proteomics, suggesting what possible secondary structures KRAS possess through bioinformatics analysis in R Studio. Our Next Generation Sequencing analysis shows the expression of KRAS mutations on primary sites, suggesting other sites than the usual might be worth investigating in the future.